requirements.txt
Flask==2.2.2
openai==0.28.0
flask_cors
mongoengine
flask_mongoengine
python-dotenv
Werkzeug==2.2.2
gunicorn